Correct asynchronous logic during lhs/rhs communication in inciter
This commit does not change the asynchronous logic in inciter, but
labels some of the LinSysMerger entry methods as [expedited]. This
appears to be required for correctness using Charm++'s randomized
message queue, which exposed this problem.

Background and details:

The symptom of the problem this commit apparently solves is a hang with
multiple CPUs during communication of the left and right hinds sides in
LinSysMerger in inciter. The problem can be reproduced using Charm++'s
randomized message queue build with

ctest -j36 -R inciter:asynclogic_pe4 -V

Printouts in LinSysMerger::addlhs() and LinSysMerger::addauxlhs()
confirm that in the above example two PEs complete addlhs(), and the two
other PEs complete addauxlhs(), but for some reason both are left
hanging never receiving the necessary messages to execute these calls
originating from the other two PEs. I have no idea why this happens,
because the two paths of execution are completely independent and are
independently waiting to be completed. For some reason, using randomized
queues, this pattern causes a deadlock. See the email thread at
https://lists.cs.illinois.edu/lists/arc/charm/2017-07/msg00022.html.

Solution:

While not very satisfactory, the problem can be avoided by labeling some
of the entry methods as [expedited], which supposedly ignore Charm++'s
randomized queue for those entry methods whose messages might be in
flight during the problematic part of the execution and thus this part
of the code execute correctly. This is what this commit does, but I
realize that ths may be simply sweeping an otherwise deeper problem
under the rug.