Skip particle output in inciter
This commit slightly changes the asynchronous logic in inciter by
commenting out the part responsible for outputting particles. While
particles in inciter are highly experimental and not tested, Charm++'s
randomized message queues exposed a problem, in which field output (not
experimental) may happen at the incorrect frequency, i.e., not at the
times the user has configured it due to particle output. This commit
fixes this problem by making the logic always skip particle output. (We
will return to fixing the particle output in some future time.)

Background and details:

Particles in inciter at this time are a highly experimental feature, not
tested at all, and by default the number of particles are zero, so most
of the code involving particle update and particle search has been
skipped already before this commit. This commit also makes the logic
skip particle file output, because it appears to be making field output
happen at incorrect times (not at times for when the user configured
it). This appears to be required for correctness using Charm++'s
randomized message queue, which exposed that due to the particle file
output logic, field output does not always happen when it should. (The
logic of particle output and field output are handled together because
at this point they are controlled by the same user configurable output
frequency. However, particle output, compared to field output, requires
more complex logic since it has to be done in multiple steps and it is
done by a library, H5Part, that writes all particle output into a single
file from multiple PEs suing MPI-IO.)