Correct ParticleWriter::writeCoords()
This is a fix, encountered while debugging the problem that led to
ebf63714.  Though ebf63714 comments out the code that this commit edits,
this fix will still most likely be required when the particle output is
fixed.

Details:

This commit puts in a correction to the logic in
ParticleWriter::writeCoords().  If the number of particles (npar) are
zero, this function is still called. When npar = 0, however,
signal2host_outcomplete() and the reduction behind it is spawned as many
times as writeCoords() is called, which is too many and incorrect. This
commit makes sure the reduction only happens once by passing in the
total number of chares which is the number of times this function is
called (once by each Carrier object) and counting up a local counter.
Only when this counter reaches the number of chares will a single signal
is started.

While this is the correct behavior for this part of overall logic,
outputting particles was entirely skipped for now by ebf63714, because
its asynchronous logic appears to be incorrect, producing erratic field
output.