Ensure that FCT data in inciter are written exclusively
This change corrects an ambiguity in the asynchronous logic in inciter
while the data, required for flux-corrected transport (FCT), is computed
and communicated, a problem exposed by Charm++'s randomized message
queues.

Background and details:

FCT requires three data structures to be computed:

(1) AEC, antidiffusive element contributions, whose sum to mesh nodes is
computed in Carrier::aec(),

(2) ALW, the maximum and minimum unknowns of elements surrounding nodes,
computed in Carrier::alw(), and

(3) LIM, the limited antidiffusive element contributions, computed in
Carrier::lim().

All the above requires communication for those mesh nodes that are on
chare boundaries. Each of the above three data structures thus consists
of two contributions, an owned and potentially multiple received ones.
Each Carrier worker object computes its own contributions to these data
and receives them from other Carrier objects. Computing the owned part
and communicating the received parts are fully overlapped in an
asynchronous fashion and are order independent but respecting the
requirement for correctness that both AEC and ALW must be available
before LIM can be computed, enforced by Charm++ SDAG logic in
src/Inciter/carrier.ci.

The owned and communicated parts of the above three data structures are
stored in Carrier's state, respectively, in { m_p, m_pc }, { m_q, m_qc
}, and {m_a, m_ac }, where the postfix 'c' stands for "communicated",
i.e., the receive buffers. While m_[pqa] and m_[pqa]c can (and are)
written to simultaneously, as the runtime system pleases to schedule the
functions that write to them, m_p must not be written to by multiple
Carrier member functions (and the same is true for m_q and m_a). For
example, m_p is written when the owned contributions to AEC are computed
and is also written when m_p and m_pc are combined, by overwriting m_p.
The same goes for m_q and m_a. The SDAG logic previously already
correctly enforced that the owned contributions to m_p and the final
combination to m_p cannot happen simultaneously. Also the runtime system
guarantees that only one member function of a chare can be in flight at
any given time, thus explicit mutexes are not needed to guard the
chare's state.

Since we don't use co-routines and Charm++ does not execute two member
functions at the same time, it appears that the current change should
not be required.  However it is required to pass the regression tests
and I cannot explain why.  Unless the logic by this commit is put in, we
get incorrect results for these data structures with randomized message
queues.