Add non-linear energy growth regression test
This is another test, whose analytical solution is known via the method
of manufactured solutions. Specification of the solution by Waltz, et.
al, "Manufactured solutions for the three-dimensional Euler equations
with relevance to Inertial Confinement Fusion", Journal of Computational
Physics 267 (2014) 196-209.

This one is time-dependent and thus exercises and tests time-dependent
Dirichlet boundary conditions.