Simplify and generalize fractional time step stage logic in inciter
This commit generalizes the number of stages per time step in inciter.
This is done by parametrizing the number of stages with n instead of
hard-coding it to be n=2 as it was previously.

The number of stages is now also exposed to user. The user can now
configure in the input file the choice of the explicit multi-stage time
stepping scheme, R_{n,0} as given by Donea, Roig, Huerta, High-order
accurate time-stepping schemes for convection-diffusion problems,
Computer Methods in Applied Mechanics and Engineering, Volume 182, Issue
3, 2000, Pages 249-275, where n is the number of stages which the user
can now configure. This enables using a different time stepping schemes
resulting in different numerical stability properties.

This commit also changes the time step stage logic. As a consequence dt
and t are now computed based on the stage in Transporter::dt() and
passed to Carrier::advance(). As a result the time step stage logic
considerably simplifies Carrier as well as some function interfaces to
FluxCorrector. This also helps prepare for correcting time-dependent
problems - which have not yet been tested.

The slotted cylinder case appears correct with these changes, but
updates to regression tests might still be required.