Implement staggered scheme to mimic Lax-Wendroff with finite elements
This is still work-in-progress. Testing with Rayleigh-Taylor both
advection and source appears to work with a new way of implementing
staggered discretization for a two-stage Lax-Wendroff time stepping with
finite elements.

Many this are in WIP state: the source interface is changing, FCT is
off, nstages and artificial viscosity will go away, inhomogeneous
Dirichlet BCs are not yet complete, most MMS problems are commented out
until their source function APIs will be changed, etc.

This will also surely require updates to regression tests.