Remove Host/Transporter from Solver's template arguments
This commit removes the Host, that has been used to template Solver.
This is a step in a series of multiple steps that attempts to untangle
some of inciter's classes in order to prepare for a new hydro scheme
using DG.

The removal of the Host template argument from Solver is facilitated by
pre-constructing Charm++ callbacks before creating the Solver Charm++
chare group and passing the such preconstructured callbacks to Solver's
constructor.  This eliminates the need for the compiler to know type
information on the host (Transporter) and thus the template argument can
be removed. Since all calls from Solver to Transporter are via
reductions, callbacks work here.  There are other, point-to-point-style
communication from Solver to Transporter, which are only used for
diagnostics: these are the m_host.pe*() entry method calls via the proxy
m_transporter. These are commented out for now - we have to see how this
functionality can be reintroduced, but if not, that's no big deal to
remove it either.

Implementation details:

In Solver::m_cb the callbacks are stored in a tagged_tuple, which helps
readability on the call-sites for the reductions by using a tag.
However, the type of m_cb is not exposed to the outside of Solver and
thus Solver's constructor takes the list of callbacks as a std::vector.
This is not great, because it is more error prone than, e.g.,
std::array, which could also be used.  However, Charm++ does not like
std::array in .ci files too much.  Another, even better, way would be to
convert the vector elements to initialize the tagged tuple by a new
overload to tagged_tuple's constructor that would take a vector which
also could perform some compile-, and runtime checks, but is left for
future work. For now the initialization of m_bc remains as lame (and as
error prone) as

      m_cb( cb[0], cb[1], cb[2], cb[3] ),

which of course presumes a static size of the vector.