Corrections to async logic to enable DG to run
DG is a stub at this point but with this commit, all the important
member functions are established that communicate with the outside world
and do the right thing so the DG can now be selected by the user and it
does time stepping (doing nothing of course).

There were some fixes necessary to the asynchronous logic as DG (as well
as CG) interacts with Transporter (and Partitioner). Both CG and DG at
this point interacts exactly the same with Transporter. CG, of course,
also interacts with solver. Those points in Transporter and Partitioner
that invoke solver are now made optional so that they are only invoked
when CG is configured. As a result, Solver is now truly optional and if
the user selects DG, the Solver Charm++ chare group is not even created.
There were also some renames in Solver, which express their functions
better.

Note that the CG/DG constructors must have the same signature as we use
them in a polymorphic fashion. As a consequence, DG also has to take the
Solver proxy even though it will not use it. Another way of doing this
is to pass the Solver proxy into CG/DG not via the constructor but via
another member function. That probably would require addition logic and
could not query the BC side sets from solver in CG:CG as it is right
now. This might chage in the future, but it is okay for now.

Also broken up Transporter's constructor to a number of slightly more
manageable pieces by delegating creating solver and partitioner to their
own functions.

Also changed some screen output lines during run to make more sense with
DG.

An attempt has also been made, for the first time, to document the
asynchronous logic among the classes Transporter, Solver, and CG using
dot via doxygen. (Thus far only the internal SDAG logic of classes have
been documentation this way.  This is in Transporter.h, following the
Transporter-local SDAG graph.  This multi-class DAG is pretty big and it
is only partial, but at least it captures some of the major interactions
among these three classes if CG is configured by the user.

The DG logic appears fine with normal as well as with randomized queues,
in serial and in parallel, with and without overdecomposition, but will
put in some asynclogic regression tests next.