Simplify the logic leading to Transporter::evaluateTime
Charm++'s randomized queue has managed to fail a single regression test
that was added in the previous commit at a single time -- I could not
reproduce it -- but with the runtime error "mis-matched client callbacks
in reduction messages". In the past this has been indicative of two
reductions in flight from the same place towards the same destination
which may be problematic.

This commit fixes that issue by simplifying the logic that leads to
Transporter::evaluateTime. It also renames it to eval(). Instead of
initiating a reduction from both the field output as well as from the
optional diagnostics output, we now simply call the field output routine
(in both CG and DG) and fall through to diagnostics. Then when
diagnostics are completed a reduction is initiated to
Transporter::diagnostics() if diagnostics are computed in a time step or
to Transporter::diagcomplete() if not. This yields a single reduction
instead of potentially two and also simplifies the SDAG logic on the
receiving side.

This commit also add a dot graph for documenting the interactions
between Transporter and DG.