Revamp/simplify interactions in inciter
The main changes are:

* The time step size are now computed with an all-reduce instead of a
reduce + broadcast in both CG and DG.

* The setup and time-stepping are now completely separated by a
reduction. This appears to be necessary, though I don't quite understand
why as it would not seem to be required for correctness. Without it
randomized-queue-build of Charm++ deadlocks some asynlogic regression
tests using CG.

* Setup is started by CG/DG::setup(), time-stepping is started by
Transporter::start(), and advancement to the next step is done by
Transporter::next() for CG, and by DG::next() for DG.

* In Transporter there is now only a single call to solver that is not
needed if not CG is configured.

* Diagnostics, field output, and one-linear status are now not
surrounded by global reductions.

* There are no global reductions during time stepping in DG.

* There is only a single reduction during time stepping in CG (to
Transporter::next).

* There is now no CG/DG::init(). The logic during setup/initialization
and leading up to and during time stepping is singificantly simplified.