Fix scrambled reductions error
The solution to the "mis-matched client callbacks in reduction messages"
in DG appears to be to execute the reduction from the constructor in two
steps: first to a group, Solver::create(), and them from there to
Transporter::comfinal() which triggers DG::setup().

In DG's main time stepping loop at this point there is still no global
synchronization: a reduction for computing diagnostics happens but
concurrently to time-stepping, i.e., when there is time and not on a
code path that the main time stepping loop would block on.

The main time stepping loop in DG at this point consists of two (empty)
steps: (1) DG::dt() invoking an all-reduce to DG::advance() computing a
minimum dt, and (2) DG::advance() which, via DG::next(), calls back to
DG::dt() again. This is the time stepping loop. At this point the only
communication among the workers is computing a global minimum of the new
time step size appearing in advance().

Note that since DG's constructor now reports back to
Transporter::comfinal() via a group, we simply use Solver from DG. Thus
Solver is now created even when DG is selected by the user, but the only
member function used, besides Solver's constructor, is
Solver::created(), which is only called from DG. Also note that this
okay, since Solver's constructor is pretty lightweight, i.e., not
initializing any heavy data, only a number of empty maps and vectors,
which is negligible.

So this logic appears to work now. Here is how this was tested: In one
terminal, using Charm++'s randomized-queue build, run

while ctest -j36 -R inciter:asynclogic; do :; done

which continuously runs (forever) all asynclogic regression tests (48)
of inciter (exercising both CG and DG) until an error triggered by a
non-zero exit code. At the same time (e.g., in another window), run

while make clean; make -sj36; do :; done

which basically does "make clean && make" continuously (forever), using
all CPUs, which creates a highly random load and a largely
oversubscribed load on all 36 CPUs. This triggeres a very random usage
of the randomized-queue asynchronous logic in Charm++, which eventually
finds the error. (Without such testing, the error could not be
triggered at all.)

Before the changes this commit puts in, sooner or later, this technique
triggered one of the parallel + overdecomposed asynclogic DG ctest tests
to fail with the "mis-matched client callbacks ..." error, usually
within roughly 10-15 seconds (after a number of times all the 48 tests
were run already). With the changes this commit puts in, I was able to
run this for several (10-15) minutes without an error, and had to
interrupt it by hand.

Well, let's see how this fares on travis and the alpine docker image
with the randomized queue build, which is the only build config that
originally triggered this problem.