Add std::get<T>(std::tuple)
This is a C++14 feature. This commit puts in an implementation that
works with C++11 and if libc++'s C++14 is detected, it uses the standard
one.

This allows reverting the default minimum language standard to C++11,
which would not necessarily be required, but makes it easier with the
TeamCity linux build agents whose default, i.e., system-wide, gcc is
4.8.5, whose libstdc++ does not yet support std::get<T>(std::tuple), for
which we use the implementation in the newly added src/Base/Get_Tuple.h.