Rename CG to MatCG in inciter
This is in preparation for #169 (enabling a diagonal, i.e., lumped,
matrix solver) for continuous Galerkin (CG) discretization as an
alternative to the existing MatCG (previously CG) which stores the left
hand side matrix in compressed sparse row storage and solves it with a
linear solver.

This commit simply renames CG to MatCG and introduces the DiagCG user
option to inciter's parser. The default discretization scheme remains
MatCG.

Also remove an unnnecessary include from Transporter.h.