Abstract away the use of FCT to DistFCT
DistFCT (short for Distributed FCT) is a new class and a Charm++ chare
array, introduced here, whose elements are bound to, i.e., migrated
together with, that of Discretization as well as MatCG and DiagCG. All
these bound arrays are created in SchemeBase's constructor. DistFCT
facilitates abstracting away all functionality and data that are related
to the use of flux corrected transport (FCT). The actual low level
details and the numerics of FCT are still done in FluxCorrector (as
before), but now FCT is used solely via DistFCT.

At this point only MatCG is hooked up with FCT but DiagCG will also be
done in a similar fashion.

Other dependent changes and implementation details:

  * DistFCT holds a number of copies of mesh-related data structures.
The original of these are in Discretization and in principle could all
be accessible if the Discretization proxy or the chare array element is
known.  However, Discretization or its proxy cannot be held in DistFCT
since that would introduce a mutual dependency between the two classes
and so that would not compile. Forward declaration is also not an
option, because we need to access member functions. When data available
from Discretization is needed in DistFCT, it is accessed  via the
Discretization const reference, see, e.g., DistFCT::aec(). This okay,
because this reference (of type Discretization) is passed via an
argument of a member function that is not a Charm++ entry method. This
would not work with an entry method, such as the constructor, because
then the Charm++ interface file would have to include DistFCT.h or
distfct.decl.h, which would introduce a cyclic dependency. Thus as long
as Discretization is passed via a non-entry method argument, this is
okay, and all of Discretization's state can be access in a given
function of DistFCT.  In member of functions of DistFCT that are not
directly called from the outside, e.g., comaec(), lim(), etc., a copy of
such data are made. At this point these are m_msum, m_bid, m_lid, and
m_inpoel. This is not ideal, but it is still cleaner than storing only a
pointer to these objects, which then would need additional code for
migrating. Another solution would have been to initialize these objects
via a member function other than the constructor but that would not be
RAII.

  * The functor "Idx" and function "element", originally in SchemeBase,
are now made generic and placed in Base/Variant.h as free templated
functor and function, respectively. This is so that all of Scheme, its
base SchemeBase, as well as the newly introduced variant,
DistFCT::m_scheme, can reuse them.  Compared to
inciter::SchemeBase::proxy, the type of DistFCT::m_scheme is a variant
that only incorporates DiagCG and MatCG Charm++ chare proxies as FCT, at
this point, is only designed to interoperate with these two schemes.
DistFCT::m_scheme facilitates type-safe callbacks to their respective
discretization scheme class objects, DiagCG and MatCG, depending on
where they are called from. This avoids having to know a single and
exclusive type (MatCG xor DiagCG) in DistFCT explicitly, but still be
able to do a callback to the right object (of the right type) in a type
safe manner. There are two such call backs at this point, in
DistFCT::verifyBC() and apply(). Another, possibly a lot messier,
solution would have been via templating DistFCT, which can of course
also achieve the xor on the types. The current solution avoids
templating the whole class DistFCT.

  * Raw object pointers to elements of Charm++ chare arrays of bound
array elements, e.g., Discretization in DiagCG, MatCG, and DG, and
DistFCT in Discretization, are now called via wrappers that contain an
assert ensuring that the pointer returned is non-null.

  * As before, FCT can be skipped in the input file by 'fct false' in
the 'discretization ... end' block of inciter, see
regression/transport/ShearDiff/shear_diffonly_nofct.q. However, since
now the use of FCT is abstracted away from the classes that use it, it
may be possible to cleanly skip FCT altogether and not just ignore its
results.  (Currently, even if FCT is turned off, it is still performed,
only its result is ignored.) Skipping FCT, however, will most likely
affect the current asynchronous logic in solver, and so this is deferred
to another day.