Solve lumped-mass diagonal system inside DiagCG
DiagCG can now solve the lumped-mass diagonal system without invoking a
(external linear) solver.

Details:
--------

* Communication on chare-boundaries among neighboring DiagCG chares
happens in a style similar to the AEC/ALW data structures assembled by
DistFCT, making use of msum (mesh chunk surrounding mesh chunk), a
hash-map that stores the list of global node IDs that a given chare
shares with (assigns to) neighbor chares.

* In DiagCG the logic during setup as well as during a time step,
leading to a new solution of the PDE system, is considerably simpler
compared to MatCG, as expected. In particular, setup only involves
computing and communicating the left-hand side (lhs) matrix, which is
now only a lumped-mass matrix, and thus is stored in a vector. Setup
does not require communicating the initial conditions since the linear
solver is a direct and not an iterative solver. The initial conditions
are set by the PDE classes as before and do not involve communication.
Once the lhs is computed and communicated, it is assumed to not change
during time stepping. During time stepping a new right hand side must be
assembled at every time step which consists of a high order right hand
side and a mass diffusion term for the low order solution. Both of these
vectors are computed and assembled on the chare boundaries independently
and asynchronously. When this is done on a given DiagCG chare, the two
solutions are handed over to DistFCT which returns the limited
antidiffusive element contributions. The logic in DiagCG contains no
global synchronization and thus potentially overlaps computation,
communication, and I/O not only during but also across time steps.

* An important change to FCT (in FluxCorrector) is that if the lhs is a
lumped-mass matrix, computing the antidiffusive element contributions
(AEC) no longer involves the high order solution increment (dUh), so
accordingly, that is now zeroed in FluxCorrector::aec() if DiagCG is
configured.

* Clarified in-code documentation at several locations.

* Solver is still used in DiagCG to access the Dirichlet boundary
conditions and to send diagnostics for output, both of which will be
moved out of Solver soon. Because of this diagnostics output does not
work with DiagCG at this point, but the solution appears correct.

Next steps:
-----------

* Diagnostics will be moved out of Solver to a new class that allows the
reuse of computing nodal diagnostics by both MatCG and DiagCG.

* The Dirichlet boundary conditions data structure will be moved out of
Solver to a new class that will be reused  by both MatCG and DiagCG.

* Add regression tests exercising DiagCG.