Fix it and t in diagnostics output in MatCG
There was a leftover from previous work that hardcoded the iteration
count and the time step for the diagnostics output. This commit fixes
that and also adds the time step size as a new column to the diagnostics
output.

Since diagnostics output is done via an optional asynchronous code path,
independent of the time stepping logic, i.e., without enforcing any
blocking on completion of the diagnostics output, it does not always
happen when it should. As a result not every time step results in a new
diagnostics line in the diag file, which is an error.

DiagCG still does not output anything except the header.