Move out side set data structure from Solver to new class
The data that is moved out of Solver by this commit is Solver::m_side,
which is now in BoundaryConditions::m_sideFileNodes. This is a map that
associates file node IDs to side set IDs for all side sets from an input
mesh file. The new class, BoundaryConditions, is a Charm++ chare group
(just like Solver), but is and will be used solely for dealing with side
sets and boundary conditions. This is currently reused by MatCG and
DiagCG but can also be used by DG and other discretizations.

BoundaryConditions is a Charm++ chare group, i.e., one per PE whose
elements never migrate, because this conveniently allows making the side
set data structure available to every PE. It is created by Transporter
(single chare on a single PE) after reading in the side set info from
the mesh file (in serial as it should be) and broadcasting the data to
every group branches via the BoundaryConditions contstructor.

At this time, a single function sideNodes() in BoundaryConditions
creates a map that assigns the local mesh node IDs (required by
discretization worker chares) to side set IDs. This is used to query and
compute the boundary conditions data structure every time step by DiagCG
and MatCG. See also next paragraph.

Another possible functionality to move to BoundaryConditions is the data
structure that stores the values (for each scalar equation solved) of
Dirichlet boundary conditions assigned to global node IDs. This is
currently in Solver::m_bc (needed by imposing boundary conditions on the
assembled matrix for MatCG) and assembled by first receiving this
boundary conditions data from workers (querying the PDEs integrated),
and computing a union of all contributions so Solver::m_bc is the same
on all PEs.