Remove aggregation of BC data from DiagCG
Not needed since the matrix is lumped and there is no need to access
off-PE BCs in (nonexistent) matrix solver.

With this commit DiagCG is now fully separated from Solver. Accordingly,
m_solver is now removed from DiagCG's state. The last call to Solver
from DiagCG remains in the constructor which is a signal that the
constructor has been called. While this signal, the reduction from
Solver:created(), can also be initiated from DiagCG, i.e., without going
through Solver, that may introduce multiple reductions in flight so that
step is left for another pass.