Reinstate synchronization in Solver::next()
This reverses the removal of the synchronization in Solver::next(), put
in by 2a127fd0. The idea was to remove the global synchronization
initiated from MatCG::eval() to Transporter::next(), which then resets
the solver for the next time step and calls MatCG::dt() for the next
step. The problem with this turned out be something like issuing too
many eval() calls continuing to the next time step and from MatCG::dt(),
calling Solver::next() multiple times, these calls probably did not
catch up with each other due to the potentially large asynchronous
overlap between time steps across chares and PEs. Note that this problem
only happened with MatCG and only when the diagnostics interval was
equal to or larger than about 5 and with more PEs than about 12. Weird
problem, hard to debug, so this commits puts back the synchronization at
the end of the time step for MatCG, irrespective of the diagnostics
interval, which appears to fix the problem.