Add forgotted doneInserting() after dynamic insertion of DistFCT chares
I'm hoping this will be a fix for a problem that came out recently
during a single regression test of about 500+ times running the full
test suites in recent days which triggered the Assert:

"DistFCT ckLocal() null"

Considering we have 80 tests for inciter, that is a single Assert
triggered out of roughly 40K inciter runs. I love asynchronous
programming! ;-)