Fix start time for MatCG/DiagCG
During the recent large revamp of the discretization classes, I put a
timer in Discretization, which has been queried when the one-liner
status is output to screen at the end of a time step. This outputs ETE,
and ETA (estimated time elapsed, and estimated time to accomplishment)
as well, so the ETE column at the end of a simulation can (and is) used
by measuring the time-stepping time. However, during the revamp this was
wrongly measuring the wall-clock time from the creation of the
Discretization class, i.e., from its constructor, which is increasingly
inaccurate for large jobs. The correct time to start this timer is when
time stepping starts, which this commit puts in for MatCG and DiagCG. DG
does not have yet this stuff.