Rename PDE to CGPDE
This commit is a first in a series that prepares the polymorphic PDE
machinery and factory to enable PDEs that use discontinuous Galerkin
discretization. The first step (this commit) is to rename the existing
PDE "base" class to CGPDE, which now will only be used as a "base" for
PDEs that use continuous Galerkin discretization and model PDE::Concept.
A new DGPDE "base" will be introduced for those PDEs that model a
DGPDE::Concept which will use discontinuous Galerkin discretization.