Add new DGPDE vector
This is similar to g_cgpde (which used to be g_pdes). The new g_dgpde
vector will hold the user-selected PDEs that use discontinuous Galerkin
discretization. Similarly to its brother, g_cgpde, g_dgpde is also
global (and const once PUPed) and distributed to all PEs as it holds
polymorphic pointers.