Register DG PDEs
This commit reorganizes the PDE/<Physics> folder by renaming some of the
files and also putting the Physics polcy classes into a cg or dg
namespace. This enables registering the DG PDEs. Note that the newly
created DGCompFlow.h, DGNavierStokes.h, and DGTransport.h at this point
simply repeat CG code, as these files were simply copied from their CG
counterpart and will be redefined in the future.

Note that so far only the Physics (and not the Problem) policies are
duplicated into CG and DG Physics policies for use with CG and DG
Transport and CompFlow. The goal is to keep it this way as the Physics
policies will significantly change depending on the discretization (CG
or DG), but the Problems policies will (hopefully) be implemented in a
way that both CG and DG physics will be able to reuse them. Note that
while this is hooked up this way, the Problem policy classes do contain
mesh-node-based code, which is special to CG. Next I will attempt to not
have to duplicate the Problem policy classes but make the current ones
generic enough instead. (If that does not work out nice, we can always
duplicate them as the Physics policies.)

Also fix a type in a unit test.