Make Transport/Problem::initialize() generic
This commit moves up the node-specific setting of initial conditions
(ICs) from the specific Transport/Problem policiy classes to
cg::Transport::initialize(). This enables writing the equivalent
elem-specific setting of ICs in dg::Transport::initialize(). From these
functions we still call down to the problem-specific evaluation of the
analytical solutions (for those problems where an analytical solution
exist) or assigning the user-specified ICs (for those problems where
this is what's approproate) via Problem::solution(). This is
accomplished by exposing this latter member function of all specific
Transport Problems as public and with the same signature, so they can be
called in a generic fashion as Problem::solution().