Make Transport/Problem::dirbc() generic
This commit moves up the node-specific query for Dirichlet boundary
conditions (BCs) from the specific Transport/Problem policiy classes to
cg::Transport::dirbc(). This enables writing the equivalent
elem-specific setting of BCs in dg::Transport::dirbc(). From these
functions we still call down to the problem-specific evaluation of the
solution increments (for those problems where an analytical solution
exist). This is accomplished by exposing this latter member function of
all specific Transport Problems as public and with the same signature,
so they can be called in a generic fashion as Problem::solinc().